lab3 second

+++ kern/kdebug.c
@@ -153,10 +153,10 @@ debuginfo_eip(uintptr_t addr, struct Eipdebuginfo *info)
 
 		// Make sure the STABS and string table memory is valid.
 		// LAB 3: Your code here.
-		if (user_mem_check(curenv, (const void*)stabs, stab_end - stabs, PTE_U)) {
+		if (user_mem_check(curenv, (const void*)stabs, (size_t)(stab_end - stabs), PTE_U)) {
             return -1;
         }
-        if (user_mem_check(curenv, (const void*)stabstr, stabstr_end - stabstr, PTE_U)) {
+        if (user_mem_check(curenv, (const void*)stabstr, (size_t)(stabstr_end - stabstr), PTE_U)) {
             return -1;
 		}
 	}
